[ZettaCamp][Module 1][Project Setup & CRUD for Student and School Data] 1.7. Implement DataLoader for School-Student Relationship: Update app.js's name (index.js) and the location (outside folder)
