+++ index.js
@@ -1,0 +1,81 @@
+// *************** IMPORT CORE ***************
+const connectDB = require('./core/database.js');
+
+// *************** IMPORT MODULE ***************
+const schoolTypeDefs = require('./modules/School/School.typeDefs.js');
+const studentTypeDefs = require('./modules/Student/Student.typeDefs.js');
+const userTypeDefs = require('./modules/User/User.typeDefs.js');
+const schoolResolvers = require('./modules/School/School.resolvers.js');
+const studentResolvers = require('./modules/Student/Student.resolvers.js');
+const userResolvers = require('./modules/User/User.resolvers.js');
+
+// *************** IMPORT LIBRARY ***************
+const express = require('express');
+const dotenv = require('dotenv');
+const { ApolloServer } = require('apollo-server-express');
+
+// *************** LOADER ***************
+const createSchoolLoader = require('./modules/School/School.loader.js');
+const createUserLoader = require('./modules/User/User.loader.js');
+const createStudentLoader = require('./modules/Student/Student.loader.js');
+
+// *************** Initialize the Express application
+const index = express();
+
+// *************** Load environment variables from .env file into process.env
+dotenv.config();
+
+// *************** Get the MONGODB_URI (Uniform Resource Identifier) value from Environment
+const MONGODB_URI = process.env.MONGODB_URI;
+
+// *************** Setup PORT Value from Environment
+const PORT = process.env.PORT || 4000;
+
+// ****************** Calling the connectDB to connect the index with the MongoDB database
+connectDB(MONGODB_URI);
+
+/**
+ * Asynchronously starts an Apollo Server instance and integrates it with the Express application.
+ *
+ * This function initializes the Apollo Server with the defined GraphQL type definitions and resolvers
+ * for schools, students, and users. It also sets up the context with DataLoaders for batching and caching
+ * database requests efficiently.
+ *
+ * The `context` provides:
+ * - `userLoader`: A DataLoader for batching user data requests.
+ * - `schoolLoader`: A DataLoader for batching school data requests.
+ * - `studentLoader`: A DataLoader for batching student data requests.
+ *
+ * @async
+ * @function startApolloServer
+ * @returns {Promise<void>} A promise that resolves once the Apollo Server is successfully started.
+ */
+async function startApolloServer() {
+  // ****************** Create a new Apollo Server instance with type definitions, resolvers, and context
+  const server = new ApolloServer({
+    typeDefs: [schoolTypeDefs, studentTypeDefs, userTypeDefs],
+    resolvers: [schoolResolvers, studentResolvers, userResolvers],
+    // ****************** Define a context function that returns a new set of DataLoaders for each request
+    context: function() {
+      return {
+        schoolLoader: createSchoolLoader(),
+        userLoader: createUserLoader(),
+        studentLoader: createStudentLoader()
+      };
+    },
+  });
+
+  // *************** Start the Apollo server
+  await server.start();
+
+  // *************** Apply Apollo middleware to Express application (index)
+  server.applyMiddleware({ index });
+
+  // *************** Start the Express server
+  index.listen(PORT, () => {
+    console.log(`Server ready at http://localhost:${PORT}${server.graphqlPath}`);
+  });
+}
+
+// *************** Call the function to launch server
+startApolloServer();
